update extension and test it in real case

+++ ext/azuredevops/setupAzd/index.ts
@@ -5,6 +5,16 @@ export async function runMain(): Promise<void> {
     try {
         task.setTaskVariable('hasRunMain', 'true');
         const os = process.platform
+        const localAppData = process.env.LocalAppData
+        const envPath = process.env.PATH
+        if (os === 'win32' && !localAppData) {
+            task.setResult(task.TaskResult.Failed, 'LocalAppData environment variable is not defined.');
+            return
+        }
+        if (!envPath) {
+            task.setResult(task.TaskResult.Failed, 'PATH environment variable is not defined.');
+            return
+        }
         const version = task.getInput('version') || 'latest'
         const windowsInstallScript = `powershell -c "$scriptPath = \\"$($env:TEMP)\\install-azd.ps1\\"; Invoke-RestMethod 'https://aka.ms/install-azd.ps1' -OutFile $scriptPath; . $scriptPath -Version '${version}' -Verbose:$true; Remove-Item $scriptPath"`
         const linuxOrMacOSInstallScript = `curl -fsSL https://aka.ms/install-azd.sh | sudo bash -s -- --version ${version} --verbose`
@@ -15,14 +25,14 @@ export async function runMain(): Promise<void> {
             console.log(cp.execSync(windowsInstallScript).toString())
             
             // Add azd to PATH
-            cp.execSync('Write-Host "##vso[task.prependpath]$($env:LocalAppData)\\Programs\\Azure Dev CLI"')
+            task.setVariable('PATH', `${envPath};${localAppData}\\Programs\\Azure Dev CLI`)
         } else {
             console.log(cp.execSync(linuxOrMacOSInstallScript).toString())
         }
 
         // Run `azd version` to make sure if azd installation failed, it returns error on windows
         if (os === 'win32') {
-            const azdVersion = `"$($env:LocalAppData)\\Programs\\Azure Dev CLI\\azd.exe" version`
+            const azdVersion = `"${localAppData}\\Programs\\Azure Dev CLI\\azd.exe" version`
             cp.execSync(azdVersion)
         }
     } catch (err: any) {
